tests/wpt_dom/tests/smoke/promise-pass.html

<!doctype html>
<meta charset="utf-8" />
<title>WPT DOM smoke: promise_test pass</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  promise_test(() => {
    return Promise.resolve(42).then((v) => {
      assert_equals(v, 42);
    });
  }, "promise_test smoke test passes");
</script>
